requirements.txt
aiogram~=3.13.1
python-decouple~=3.8